requirements.txt
celery
django
djangorestframework
django-url-filter
psycopg2
gunicorn
pytz
raven
redis